Move model tests to own package `tests/model`
Also, put the `gerok` and `sample` tests into their own module.
